requirements.txt
# Bank Management System - Requirements

## Python Dependencies

mysql-connector-python>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0
reportlab>=3.6.0

## Optional Dependencies (for enhanced features)

pillow>=8.0.0          # For image processing
matplotlib>=3.5.0      # For charts and graphs
seaborn>=0.11.0        # For statistical visualizations
tkinter-tooltip>=2.0.0 # For enhanced tooltips

## Development Dependencies

pytest>=6.0.0          # For testing
pytest-cov>=2.12.0     # For test coverage
black>=21.0.0          # For code formatting
flake8>=4.0.0          # For code linting
mypy>=0.910            # For type checking

## Installation Instructions

1. Install Python 3.8 or higher
2. Install MySQL Server 8.0 or higher
3. Run: pip install -r requirements.txt
4. Update config.py with your database credentials
5. Execute database/schema.sql in MySQL
6. Execute database/sample_data.sql in MySQL
7. Run: python main.py

## System Requirements

- Python 3.8+
- MySQL 8.0+
- 4GB RAM minimum
- 500MB disk space
- Windows 10/macOS 10.15/Ubuntu 18.04 or later
